+++ packages/common/src/Components/LinkPanelContainer.jsx
@@ -22,34 +22,31 @@ class LinkPanelContainer extends PureComponent {
   constructor(props) {
     super(props);
     this.styles = mergeStyles({ styles, theme: props.theme });
+    const { url, targetBlank, nofollow } = this.props;
+    this.state = {
+      linkPanelValues: { url, targetBlank, nofollow },
+    };
   }
 
-  setLinkPanel = linkPanel => this.linkPanel = linkPanel;
-
-  onDoneClick = () => {
-    const { onCancel, onDelete } = this.props;
-    if (this.linkPanel.state.isValidUrl && this.linkPanel.state.url) {
-      const { url, targetBlank, nofollow } = this.linkPanel.state;
-      this.props.onDone && this.props.onDone({ url, targetBlank, nofollow });
-    } else if (this.linkPanel.state.intermediateUrl === '') {
-      onDelete();
-      onCancel();
-    } else {
-      this.linkPanel.validateUrl();
+  onDone = () => {
+    const { linkPanelValues } = this.state;
+    if (linkPanelValues.isValid && linkPanelValues.url) {
+      this.props.onDone(linkPanelValues);
+    } else if (linkPanelValues.url === '') {
+      this.onDelete();
     }
   };
 
-  onCancelClick = () => this.props.onCancel && this.props.onCancel();
+  onDelete = () => {
+    this.props.onDelete();
+    this.onCancel();
+  };
 
-  onDeleteClick = () => {
-    const { onCancel, onDelete } = this.props;
-    onDelete();
-    onCancel();
-  }
+  onCancel = () => this.props.onCancel();
 
   render() {
     const { styles } = this;
-    const { url, targetBlank, nofollow, theme, isActive, anchorTarget, relValue, isMobile, t, ariaProps, tabIndex, uiSettings } = this.props;
+    const { theme, isActive, anchorTarget, relValue, isMobile, t, ariaProps, tabIndex, uiSettings } = this.props;
     const doneButtonText = t('LinkPanelContainer_DoneButton');
     const cancelButtonText = t('LinkPanelContainer_CancelButton');
     const removeButtonText = t('LinkPanelContainer_RemoveButton');
@@ -66,9 +63,12 @@ class LinkPanelContainer extends PureComponent {
       <FocusManager className={linkPanelContainerClassName} data-hook="linkPanelContainer" role="form" {...ariaProps}>
         <div className={styles.linkPanel_content}>
           <LinkPanel
-            onEnter={this.onDoneClick.bind(this)} onEscape={this.onCancelClick.bind(this)}
-            ref={this.setLinkPanel} theme={theme} url={url} targetBlank={targetBlank} anchorTarget={anchorTarget}
-            relValue={relValue} nofollow={nofollow} t={t} ariaProps={linkPanelAriaProps} uiSettings={uiSettings}
+            onEnter={this.onDone} onEscape={this.onCancel}
+            linkValues={this.state.linkPanelValues} onChange={linkPanelValues => this.setState({ linkPanelValues })}
+            theme={theme}
+            showTargetBlankCheckbox={uiSettings.linkPanel.blankTargetToggleVisibilityFn(anchorTarget)}
+            showRelValueCheckbox={uiSettings.linkPanel.nofollowRelToggleVisibilityFn(relValue)}
+            t={t} ariaProps={linkPanelAriaProps} {...uiSettings.linkPanel}
           />
           <div className={styles.linkPanel_actionsDivider} role="separator" />
         </div>
@@ -76,7 +76,7 @@ class LinkPanelContainer extends PureComponent {
           <div className={styles.linkPanel_LeftActions}>
             <button
               tabIndex={tabIndex} aria-label={cancelButtonText}
-              className={cancelButtonClassName} data-hook="linkPanelContainerCancel" onClick={this.onCancelClick}
+              className={cancelButtonClassName} data-hook="linkPanelContainerCancel" onClick={this.onCancel}
             >{cancelButtonText}
             </button>
             {isActive &&
@@ -85,7 +85,7 @@ class LinkPanelContainer extends PureComponent {
               <button
                 tabIndex={tabIndex} aria-label={removeButtonText}
                 className={removeButtonClassName}
-                data-hook="linkPanelContainerRemove" onClick={this.onDeleteClick}
+                data-hook="linkPanelContainerRemove" onClick={this.onDelete}
               >{removeButtonText}
               </button>
             </div>
@@ -93,7 +93,7 @@ class LinkPanelContainer extends PureComponent {
           </div>
           <button
             tabIndex={tabIndex} aria-label={doneButtonText}
-            className={doneButtonClassName} data-hook="linkPanelContainerDone" onClick={this.onDoneClick}
+            className={doneButtonClassName} data-hook="linkPanelContainerDone" onClick={this.onDone}
           >{doneButtonText}
           </button>
         </div>